Refactor chatwidget settings surfaces into modules (phase 4)
